Avoids creating a new price every time

+++ migrations/alembic/versions/3470a90fb575_create_usage_table.py
@@ -28,7 +28,7 @@ def upgrade() -> None:
         "usage",
         sa.Column("id", sa.BigInteger, id_seq, server_default=id_seq.next_value(), nullable=False, primary_key=True),
         sa.Column("usage_date", sa.Date, nullable=False, server_default=sa.func.current_date()),
-        sa.Column("kwh_consumed", sa.Float, nullable=False),
+        sa.Column("kwh_consumed", sa.Integer, nullable=False),
         sa.Column("device_id", sa.BigInteger, nullable=False),
         sa.Column("created_at", sa.DateTime, nullable=False, server_default=sa.func.current_timestamp()),
         sa.Column("updated_at", sa.DateTime, nullable=False, server_default=sa.func.current_timestamp()),
